:sparkles: adicionando tratamento de erro para id de usuário inexistente

+++ backend/src/services/user.service.js
@@ -11,7 +11,11 @@ module.exports = {
   },
   getUserById: async (id) => {
     try {
-      return { error: null, output: await User.findByPk(id) };
+      const user = await User.findByPk(id);
+
+      if (!user) return { error: 'NOT_FOUND', output: 'User not found.' };
+
+      return { error: null, output: user };
     } catch (error) {
       console.error(error);
       return { error: 'INTERNAL_ERROR' };
@@ -29,6 +33,10 @@ module.exports = {
 
   deleteUser: async (id) => {
     try {
+      const user = await User.findByPk(id);
+
+      if (!user) return { error: 'NOT_FOUND', output: 'User not found.' };
+
       await User.destroy({ where: { id } });
 
       return { error: null };
